Extracted services from Program.cs, implemented test measure API
Moved services into the Services folder and interfaces into their own file. We should look into what best practice for the massive service is going to be for handling the actual database.

Also implemented a slight test API for measurements to confirm enum mapping is working, it is!
